fix for dataproc

+++ http/src/main/resources/init-resources/startup.sh
@@ -96,12 +96,14 @@ else
     GCLOUD_CMD='gcloud'
     DOCKER_COMPOSE='docker-compose'
 
-    # Update old welder docker-compose file's entrypoint
-    sed -i  "s/tail -f \/dev\/null/\/opt\/docker\/bin\/entrypoint.sh/g" /etc/welder-docker-compose.yaml
+    if [ "$WELDER_ENABLED" == "true" ] ; then
+      # Update old welder docker-compose file's entrypoint
+      sed -i  "s/tail -f \/dev\/null/\/opt\/docker\/bin\/entrypoint.sh/g" /etc/welder-docker-compose.yaml
 
-    ${DOCKER_COMPOSE} -f ${WELDER_DOCKER_COMPOSE} stop
-    ${DOCKER_COMPOSE} -f ${WELDER_DOCKER_COMPOSE} rm -f
-    ${DOCKER_COMPOSE} -f ${WELDER_DOCKER_COMPOSE} up -d &> /var/start_output.txt || EXIT_CODE=$?
+      ${DOCKER_COMPOSE} -f ${WELDER_DOCKER_COMPOSE} stop
+      ${DOCKER_COMPOSE} -f ${WELDER_DOCKER_COMPOSE} rm -f
+      ${DOCKER_COMPOSE} -f ${WELDER_DOCKER_COMPOSE} up -d &> /var/start_output.txt || EXIT_CODE=$?
+    fi
 fi
 
 function failScriptIfError() {
